requirements.txt
# correlation_cuda==0.0.0
# csrc==0.0.0
# geatpy==2.7.0
joblib==1.4.0
# mamba_ssm==2.1.0
matplotlib==3.5.3
numpy==1.24.3
nuscenes_devkit==1.1.11
open3d==0.18.0
opencv_python==4.9.0.80
pandas==2.0.3
Pillow==11.0.0
problem==0.2.3
pykitti==0.3.1
pyquaternion==0.9.9
# PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.3.2
scipy==1.14.1
setuptools==68.2.2
spconv_cu116==2.3.6
timm==1.0.11
torch==1.13.1
torchinfo==1.8.0
torchvision==0.14.1
tqdm==4.66.2
